requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.3.3
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.1.1
conda==4.7.12
conda-package-handling==1.3.11
cryptography==2.7
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
filelock==3.0.12
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
image==1.5.27
importlib-metadata==1.5.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
libarchive-c==2.8
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.2.1
pipenv==2018.11.26
psycopg2==2.8.5
pycosat==0.6.3
pycparser==2.19
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyOpenSSL==19.0.0
pyotp==2.3.0
PySocks==1.7.0
pytz==2019.2
qrcode==6.1
requests==2.22.0
ruamel-yaml==0.15.46
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
tqdm==4.32.1
typed-ast==1.4.1
urllib3==1.24.2
virtualenv==20.0.14
virtualenv-clone==0.5.4
Werkzeug==1.0.0
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
